async removed, code clean up for sync, with modif, to do perftests
