+++ DevFast.Net.Text/docs/DevFast.Net.Text.Json.Utf8.md
@@ -4,5 +4,5 @@
 
 | Classes | |
 | :--- | :--- |
-| [AsyncUtf8JsonArrayPartReader](DevFast.Net.Text.Json.Utf8.AsyncUtf8JsonArrayPartReader.md 'DevFast.Net.Text.Json.Utf8.AsyncUtf8JsonArrayPartReader') | Class implementing [IAsyncJsonArrayPartReader](DevFast.Net.Text.Json.IAsyncJsonArrayPartReader.md 'DevFast.Net.Text.Json.IAsyncJsonArrayPartReader') for standard Utf-8 JSON data encoding based on https://datatracker.ietf.org/doc/html/rfc7159 (grammar shown at https://www.json.org/json-en.html).   This implementation support both single line comments (starting with '//' and ending in either Carriage return '\r' or newline '\n') and multiline comments (starting with '/*' and ending with '*/'). |
-| [AsyncUtf8MemJsonArrayPartReader](DevFast.Net.Text.Json.Utf8.AsyncUtf8MemJsonArrayPartReader.md 'DevFast.Net.Text.Json.Utf8.AsyncUtf8MemJsonArrayPartReader') | Class implementing [IAsyncJsonArrayPartReader](DevFast.Net.Text.Json.IAsyncJsonArrayPartReader.md 'DevFast.Net.Text.Json.IAsyncJsonArrayPartReader') for standard Utf-8 JSON data encoding based on https://datatracker.ietf.org/doc/html/rfc7159 (grammar shown at https://www.json.org/json-en.html).   This implementation support both single line comments (starting with '//' and ending in either Carriage return '\r' or newline '\n') and multiline comments (starting with '/*' and ending with '*/'). |
+| [JsonArrayReader](DevFast.Net.Text.Json.Utf8.JsonArrayReader.md 'DevFast.Net.Text.Json.Utf8.JsonArrayReader') | Class implementing [IJsonArrayReader](DevFast.Net.Text.Json.IJsonArrayReader.md 'DevFast.Net.Text.Json.IJsonArrayReader') for standard Utf-8 JSON data encoding based on https://datatracker.ietf.org/doc/html/rfc7159 (grammar shown at https://www.json.org/json-en.html).   This implementation support both single line comments (starting with '//' and ending in either Carriage return '\r' or newline '\n') and multiline comments (starting with '/*' and ending with '*/'). |
+| [MemJsonArrayReader](DevFast.Net.Text.Json.Utf8.MemJsonArrayReader.md 'DevFast.Net.Text.Json.Utf8.MemJsonArrayReader') | Class implementing [IJsonArrayReader](DevFast.Net.Text.Json.IJsonArrayReader.md 'DevFast.Net.Text.Json.IJsonArrayReader') for standard Utf-8 JSON data encoding based on https://datatracker.ietf.org/doc/html/rfc7159 (grammar shown at https://www.json.org/json-en.html).   This implementation support both single line comments (starting with '//' and ending in either Carriage return '\r' or newline '\n') and multiline comments (starting with '/*' and ending with '*/'). |
