+++ DevFast.Net.Text/src/DevFast.Net.Text/Json/IJsonArrayReader.cs
@@ -1,15 +1,15 @@
 ﻿namespace DevFast.Net.Text.Json
 {
     /// <summary>
-    /// Interface dictating implementation to asynchronously parse individual items of a JSON Array 
-    /// or to parse individual elements (as defined in <see cref="JsonType"/>) in a JSON sequence.
+    /// Interface dictating implementation of parsing individual items of a JSON Array 
+    /// with the possibility to parse individual elements (as defined in <see cref="JsonType"/>) in a JSON sequence.
     /// Parsing of such elements produces <see cref="RawJson"/> representing entire value-form 
     /// (including structural characters, string quotes etc.) as <see cref="RawJson.Value"/>, of single element at a time, 
     /// of a known <see cref="JsonType"/>.
-    /// Implementation should avoid any kind of deserialization and, for that, should use other standard
+    /// For the moment implementation should avoid any kind of deserialization and, for that, should use other standard
     /// JSON deserialization libraries.
     /// </summary>
-    public interface IAsyncJsonArrayPartReader : IDisposable
+    public interface IJsonArrayReader : IDisposable
     {
         /// <summary>
         /// <see langword="true"/> indicating that reader has reached end of JSON input,
@@ -34,7 +34,7 @@
         int Capacity { get; }
 
         /// <summary>
-        /// Provides a convenient way to asynchronously enumerate over elements of a JSON array (one at a time).
+        /// Provides a convenient way to enumerate over elements of a JSON array (one at a time).
         /// For every iteration, such mechanism produces <see cref="RawJson"/>, where <see cref="RawJson.Value"/> represents
         /// entire value-form (including structural characters, string quotes etc.) of such an individual
         /// element &amp; <see cref="RawJson.Type"/> indicates underlying JSON element type. 
@@ -46,17 +46,7 @@
         /// single line and multiline comments are allowed after <see cref="JsonConst.ArrayEndByte"/> until <see cref="EoJ"/>.</param>
         /// <param name="token">Cancellation token to observe.</param>
         /// <exception cref="JsonArrayPartParsingException"></exception>
-        IEnumerable<RawJson> EnumerateRawJsonArrayElementAsync(bool ensureEoj, CancellationToken token = default);
-
-        /// <summary>
-        /// Call makes reader skip all the irrelevant whitespaces (comments included). Once done, it returns
-        /// <see langword="true"/> if value is <see cref="JsonConst.ArrayBeginByte"/>. If the value matches, 
-        /// then reader advances its current position to next <see cref="byte"/> in the sequence or to end of JSON.
-        /// Otherwise, it returns <see langword="false"/> when current byte is NOT <see cref="JsonConst.ArrayBeginByte"/> and
-        /// reader position is maintained on the current byte.
-        /// </summary>
-        /// <param name="token">Cancellation token to observe</param>
-        bool ReadIsBeginArrayAsync(CancellationToken token = default);
+        IEnumerable<RawJson> EnumerateJsonArray(bool ensureEoj, CancellationToken token = default);
 
         /// <summary>
         /// Call makes reader skip all the irrelevant whitespaces (comments included). Once done, it checks
@@ -67,7 +57,17 @@
         /// </summary>
         /// <param name="token">Cancellation token to observe</param>
         /// <exception cref="JsonArrayPartParsingException"></exception>
-        void ReadIsBeginArrayWithVerifyAsync(CancellationToken token = default);
+        void ReadIsBeginArrayWithVerify(CancellationToken token = default);
+
+        /// <summary>
+        /// Call makes reader skip all the irrelevant whitespaces (comments included). Once done, it returns
+        /// <see langword="true"/> if value is <see cref="JsonConst.ArrayBeginByte"/>. If the value matches, 
+        /// then reader advances its current position to next <see cref="byte"/> in the sequence or to end of JSON.
+        /// Otherwise, it returns <see langword="false"/> when current byte is NOT <see cref="JsonConst.ArrayBeginByte"/> and
+        /// reader position is maintained on the current byte.
+        /// </summary>
+        /// <param name="token">Cancellation token to observe</param>
+        bool ReadIsBeginArray(CancellationToken token = default);
 
         /// <summary>
         /// Call makes reader skip all the irrelevant whitespaces (comments included). Once done, it returns
@@ -81,15 +81,14 @@
         /// <see langword="true"/> to ensure that no data is present after <see cref="JsonConst.ArrayEndByte"/>. However, both
         /// single line and multiline comments are allowed before <see cref="EoJ"/>.</param>
         /// <param name="token">Cancellation token to observe</param>
-        bool ReadIsEndArrayAsync(bool ensureEoj, CancellationToken token = default);
+        bool ReadIsEndArray(bool ensureEoj, CancellationToken token = default);
 
         /// <summary>
-        /// Reads the current JSON element as <see cref="RawJson"/>. If reaches <see cref="EoJ"/> or
-        /// encounters <see cref="JsonConst.ArrayEndByte"/>, returned <see cref="RawJson.Type"/> is
-        /// <see cref="JsonType.Nothing"/>.
+        /// Reads the current JSON element as <see cref="RawJson"/>. If it reaches <see cref="EoJ"/> or
+        /// encounters <see cref="JsonConst.ArrayEndByte"/>, it returns <see cref="JsonType.Nothing"/> as
+        /// <see cref="RawJson.Type"/>.
         /// <para>
-        /// One should prefer <see cref="EnumerateRawJsonArrayElementAsync(bool, CancellationToken)"/>
-        /// to parse well-structured JSON stream over this method.
+        /// One should prefer <see cref="EnumerateJsonArray"/> to parse well-structured JSON stream over this method.
         /// This method is to parse non-standard chain of JSON elements separated by ',' (or not).
         /// </para>
         /// </summary>
@@ -97,6 +96,6 @@
         /// after successfully parsing the current JSON element; <see langword="false"/> otherwise.</param>
         /// <param name="token">Cancellation token to observe.</param>
         /// <exception cref="JsonArrayPartParsingException"></exception>
-        RawJson GetCurrentRawAsync(bool withVerify = true, CancellationToken token = default);
+        RawJson ReadRaw(bool withVerify = true, CancellationToken token = default);
     }
 }
