+++ DevFast.Net.Text/src/DevFast.Net.Text/Json/JsonReader.cs
@@ -10,16 +10,16 @@ namespace DevFast.Net.Text.Json
     public static class JsonReader
     {
         /// <summary>
-        /// Convenient method, to create well initialized instance of type <see cref="IAsyncJsonArrayPartReader"/>.
+        /// Convenient method, to create well initialized instance of type <see cref="IJsonArrayReader"/>.
         /// </summary>
         /// <param name="stream">A readable stream containing JSON array data.</param>
         /// <param name="token">Cancellation token to observe.</param>
         /// <param name="size">Initial size of underlying byte buffer. Any value less than
         /// <see cref="JsonConst.RawUtf8JsonPartReaderMinBuffer"/> will be ignored.</param>
         /// <param name="disposeStream"><see langword="true"/> to dispose <paramref name="stream"/> when either
-        /// current instance of <see cref="IAsyncJsonArrayPartReader"/> itself disposing or
+        /// current instance of <see cref="IJsonArrayReader"/> itself disposing or
         /// when <paramref name="stream"/> is completely read; <see langword="false"/> otherwise.</param>
-        public static async ValueTask<IAsyncJsonArrayPartReader> CreateAsync(Stream stream,
+        public static async ValueTask<IJsonArrayReader> CreateAsync(Stream stream,
             CancellationToken token = default,
             int size = JsonConst.RawUtf8JsonPartReaderMinBuffer,
             bool disposeStream = false)
@@ -40,7 +40,7 @@ namespace DevFast.Net.Text.Json
                 {
                     newOffSet = 3;
                 }
-                return new AsyncUtf8MemJsonArrayPartReader(ms, 
+                return new MemJsonArrayReader(ms, 
                     new ArraySegment<byte>(segment.Array!, segment.Offset + newOffSet, segment.Count - newOffSet), 
                     disposeStream);
             }
@@ -65,7 +65,7 @@ namespace DevFast.Net.Text.Json
                 {
                     begin = 3;
                 }
-                return new AsyncUtf8JsonArrayPartReader(stream, buffer, begin, end, disposeStream);
+                return new JsonArrayReader(stream, buffer, begin, end, disposeStream);
             }
         }
