+++ DevFast.Net.Text/src/DevFast.Net.Text/Json/Utf8/JsonArrayReader.cs
@@ -1,30 +1,33 @@
-﻿using System.Runtime.CompilerServices;
-using DevFast.Net.Extensions.SystemTypes;
+﻿using DevFast.Net.Extensions.SystemTypes;
+using System;
+using System.Runtime.CompilerServices;
+using System.Text;
 
 namespace DevFast.Net.Text.Json.Utf8
 {
     /// <summary>
-    /// Class implementing <see cref="IAsyncJsonArrayPartReader"/> for standard Utf-8 JSON data encoding
+    /// Class implementing <see cref="IJsonArrayReader"/> for standard Utf-8 JSON data encoding
     /// based on https://datatracker.ietf.org/doc/html/rfc7159 (grammar shown at https://www.json.org/json-en.html).
     /// <para>
     /// This implementation support both single line comments (starting with '//' and ending in either Carriage return '\r'
     /// or newline '\n') and multiline comments (starting with '/*' and ending with '*/').
     /// </para>
     /// </summary>
-    public sealed class AsyncUtf8MemJsonArrayPartReader : IAsyncJsonArrayPartReader
+    public sealed class JsonArrayReader : IJsonArrayReader
     {
         private readonly bool _disposeStream;
-        private readonly int _end;
-        private MemoryStream? _stream;
+        private Stream? _stream;
         private byte[] _buffer;
-        private int _current;
+        private int _begin, _end, _current;
+        private long _bytesConsumed;
 
-        internal AsyncUtf8MemJsonArrayPartReader(MemoryStream stream, ArraySegment<byte> internalSegment, bool disposeStream)
+        internal JsonArrayReader(Stream stream, byte[] buffer, int begin, int end, bool disposeStream)
         {
             _stream = stream;
-            _buffer = internalSegment.Array!;
-            _current = internalSegment.Offset;
-            _end = internalSegment.Offset + internalSegment.Count;
+            _buffer = buffer;
+            _current = _begin = begin;
+            _bytesConsumed = _begin;
+            _end = end;
             _disposeStream = disposeStream;
         }
 
@@ -32,7 +35,7 @@ namespace DevFast.Net.Text.Json.Utf8
         /// <see langword="true"/> indicating that reader has reached end of JSON input,
         /// otherwise <see langword="false"/>.
         /// </summary>
-        public bool EoJ => !InRange;
+        public bool EoJ => _stream == null && _current == _end;
 
         /// <summary>
         /// <see cref="byte"/> value of current position of reader. <see langword="null"/> when
@@ -43,17 +46,15 @@ namespace DevFast.Net.Text.Json.Utf8
         /// <summary>
         /// Total number of <see cref="byte"/>s observed by the reader since the very beginning (0-based position).
         /// </summary>
-        public long Position => Math.Min(_current, _end);
+        public long Position => _bytesConsumed + (_current - _begin);
 
         /// <summary>
         /// Current capacity as total number of <see cref="byte"/>s.
         /// </summary>
         public int Capacity => _buffer.Length;
 
-        private bool InRange => _current < _end;
-
         /// <summary>
-        /// Provides a convenient way to asynchronously enumerate over elements of a JSON array (one at a time).
+        /// Provides a convenient way to enumerate over elements of a JSON array (one at a time).
         /// For every iteration, such mechanism produces <see cref="RawJson"/>, where <see cref="RawJson.Value"/> represents
         /// entire value-form (including structural characters, string quotes etc.) of such an individual
         /// element &amp; <see cref="RawJson.Type"/> indicates underlying JSON element type. 
@@ -65,17 +66,17 @@ namespace DevFast.Net.Text.Json.Utf8
         /// single line and multiline comments are allowed after <see cref="JsonConst.ArrayEndByte"/> until <see cref="EoJ"/>.</param>
         /// <param name="token">Cancellation token to observe.</param>
         /// <exception cref="JsonArrayPartParsingException"></exception>
-        public IEnumerable<RawJson> EnumerateRawJsonArrayElementAsync(bool ensureEoj,
+        public IEnumerable<RawJson> EnumerateJsonArray(bool ensureEoj,
             CancellationToken token = default)
         {
-            ReadIsBeginArrayWithVerifyAsync(token);
-            while (!ReadIsEndArrayAsync(ensureEoj, token))
+            ReadIsBeginArrayWithVerify(token);
+            while (!ReadIsEndArray(ensureEoj, token))
             {
-                var next = GetCurrentRawAsync(true, token);
+                var next = ReadRaw(true, token);
                 if (next.Type == JsonType.Nothing)
                 {
                     throw new JsonArrayPartParsingException($"Expected a valid JSON element or end of JSON array. " +
-                                                            $"0-Based Position = {Position}.");
+                        $"0-Based Position = {Position}.");
                 }
                 yield return next;
             }
@@ -90,22 +91,22 @@ namespace DevFast.Net.Text.Json.Utf8
         /// </summary>
         /// <param name="token">Cancellation token to observe</param>
         /// <exception cref="JsonArrayPartParsingException"></exception>
-        public void ReadIsBeginArrayWithVerifyAsync(CancellationToken token = default)
+        public void ReadIsBeginArrayWithVerify(CancellationToken token = default)
         {
-            if (!ReadIsBeginArrayAsync(token))
+            if (!ReadIsBeginArray(token))
             {
                 if (InRange)
                 {
                     throw new JsonArrayPartParsingException("Invalid byte value for JSON begin-array. " +
                                                             $"Expected = {JsonConst.ArrayBeginByte}, " +
-                                                            $"Found = {_buffer[_current]}, " +
+                                                            $"Found = {(char)Current!}, " +
                                                             $"0-Based Position = {Position}.");
                 }
                 throw new JsonArrayPartParsingException("Reached end, unable to find JSON begin-array." +
                                                         $"0-Based Position = {Position}.");
             }
         }
-        
+
         /// <summary>
         /// Call makes reader skip all the irrelevant whitespaces (comments included). Once done, it returns
         /// <see langword="true"/> if value is <see cref="JsonConst.ArrayBeginByte"/>. If the value matches, 
@@ -114,9 +115,9 @@ namespace DevFast.Net.Text.Json.Utf8
         /// reader position is maintained on the current byte.
         /// </summary>
         /// <param name="token">Cancellation token to observe</param>
-        public bool ReadIsBeginArrayAsync(CancellationToken token = default)
+        public bool ReadIsBeginArray(CancellationToken token = default)
         {
-            return ReadIsGivenByte(JsonConst.ArrayBeginByte, token);
+            return ReadIsGivenByteAsync(JsonConst.ArrayBeginByte, token);
         }
 
         /// <summary>
@@ -131,14 +132,14 @@ namespace DevFast.Net.Text.Json.Utf8
         /// <see langword="true"/> to ensure that no data is present after <see cref="JsonConst.ArrayEndByte"/>. However, both
         /// single line and multiline comments are allowed before <see cref="EoJ"/>.</param>
         /// <param name="token">Cancellation token to observe</param>
-        public bool ReadIsEndArrayAsync(bool ensureEoj, CancellationToken token = default)
+        public bool ReadIsEndArray(bool ensureEoj, CancellationToken token = default)
         {
-            var reply = ReadIsGivenByte(JsonConst.ArrayEndByte, token);
-            if (ensureEoj && reply)
+            var reply = ReadIsGivenByteAsync(JsonConst.ArrayEndByte, token);
+            if(ensureEoj && reply)
             {
                 //we need to make sure only comments exists or we reached EOJ!
-                SkipWhiteSpace();
-                if (!EoJ)
+                SkipWhiteSpaceAsync(token);
+                if(!EoJ)
                 {
                     throw new JsonArrayPartParsingException($"Expected End Of JSON after encountering ']'. " +
                                                             $"0-Based Position = {Position}.");
@@ -152,7 +153,7 @@ namespace DevFast.Net.Text.Json.Utf8
         /// encounters <see cref="JsonConst.ArrayEndByte"/>, returned <see cref="RawJson.Type"/> is
         /// <see cref="JsonType.Nothing"/>.
         /// <para>
-        /// One should prefer <see cref="EnumerateRawJsonArrayElementAsync(bool, CancellationToken)"/>
+        /// One should prefer <see cref="EnumerateJsonArray"/>
         /// to parse well-structured JSON stream over this method.
         /// This method is to parse non-standard chain of JSON elements separated by ',' (or not).
         /// </para>
@@ -161,120 +162,143 @@ namespace DevFast.Net.Text.Json.Utf8
         /// after successfully parsing the current JSON element; <see langword="false"/> otherwise.</param>
         /// <param name="token">Cancellation token to observe.</param>
         /// <exception cref="JsonArrayPartParsingException"></exception>
-        public RawJson GetCurrentRawAsync(bool withVerify = true, CancellationToken token = default)
+        public RawJson ReadRaw(bool withVerify = true, CancellationToken token = default)
         {
-            SkipWhiteSpace();
-            if (!InRange || _buffer[_current] == JsonConst.ArrayEndByte) return new RawJson(JsonType.Nothing, Array.Empty<byte>());
-            var begin = _current;
-            var type = SkipUntilNextRaw();
-            token.ThrowIfCancellationRequested();
-            var currentRaw = new byte[_current - begin];
-            _buffer.CopyToUnSafe(currentRaw, begin, currentRaw.Length, 0);
+            SkipWhiteSpaceAsync(token);
+            if (!InRange || _buffer.Current == JsonConst.ArrayEndByte) return new RawJson(JsonType.Nothing, Array.Empty<byte>());
+            _buffer.SkipUntilCurrent();
+            var type = SkipUntilNextRawAsync(token);
+            var currentRaw = _buffer.GetUntilCurrent();
 
-            if (withVerify)
+            if(withVerify)
             {
-                ReadIsValueSeparationOrEndWithVerify(JsonConst.ArrayEndByte,
+                ReadIsValueSeparationOrEndWithVerifyAsync(JsonConst.ArrayEndByte,
                         "array",
                         "',' or ']' (but not ',]')",
                         token);
             }
             else
             {
-                ReadIsGivenByte(JsonConst.ValueSeparatorByte, token);
+                ReadIsGivenByteAsync(JsonConst.ValueSeparatorByte, token);
             }
             return new RawJson(type, currentRaw);
         }
 
-        private JsonType SkipUntilNextRaw()
+        private JsonType SkipUntilNextRawAsync(CancellationToken token)
         {
-            return _buffer[_current] switch
+            return _buffer.Current switch
             {
-                JsonConst.ArrayBeginByte => SkipArray(),
-                JsonConst.ObjectBeginByte => SkipObject(),
-                JsonConst.StringQuoteByte => SkipString(),
-                JsonConst.MinusSignByte => SkipNumberWithoutValidation(),
-                JsonConst.Number0Byte => SkipNumberWithoutValidation(),
-                JsonConst.Number1Byte => SkipNumberWithoutValidation(),
-                JsonConst.Number2Byte => SkipNumberWithoutValidation(),
-                JsonConst.Number3Byte => SkipNumberWithoutValidation(),
-                JsonConst.Number4Byte => SkipNumberWithoutValidation(),
-                JsonConst.Number5Byte => SkipNumberWithoutValidation(),
-                JsonConst.Number6Byte => SkipNumberWithoutValidation(),
-                JsonConst.Number7Byte => SkipNumberWithoutValidation(),
-                JsonConst.Number8Byte => SkipNumberWithoutValidation(),
-                JsonConst.Number9Byte => SkipNumberWithoutValidation(),
-                JsonConst.FirstOfTrueByte => SkipRueOfTrueWithoutValidation(),
-                JsonConst.FirstOfFalseByte => SkipAlseOfFalseWithoutValidation(),
-                JsonConst.FirstOfNullByte => SkipUllOfNullWithoutValidation(),
+                JsonConst.ArrayBeginByte => SkipArrayAsync(token),
+                JsonConst.ObjectBeginByte => SkipObjectAsync(token),
+                JsonConst.StringQuoteByte => SkipStringAsync(token),
+                JsonConst.MinusSignByte => SkipNumberWithoutValidationAsync(token),
+                JsonConst.Number0Byte => SkipNumberWithoutValidationAsync(token),
+                JsonConst.Number1Byte => SkipNumberWithoutValidationAsync(token),
+                JsonConst.Number2Byte => SkipNumberWithoutValidationAsync(token),
+                JsonConst.Number3Byte => SkipNumberWithoutValidationAsync(token),
+                JsonConst.Number4Byte => SkipNumberWithoutValidationAsync(token),
+                JsonConst.Number5Byte => SkipNumberWithoutValidationAsync(token),
+                JsonConst.Number6Byte => SkipNumberWithoutValidationAsync(token),
+                JsonConst.Number7Byte => SkipNumberWithoutValidationAsync(token),
+                JsonConst.Number8Byte => SkipNumberWithoutValidationAsync(token),
+                JsonConst.Number9Byte => SkipNumberWithoutValidationAsync(token),
+                JsonConst.FirstOfTrueByte => SkipRueOfTrueWithoutValidationAsync(token),
+                JsonConst.FirstOfFalseByte => SkipAlseOfFalseWithoutValidationAsync(token),
+                JsonConst.FirstOfNullByte => SkipUllOfNullWithoutValidationAsync(token),
                 _ => throw new JsonArrayPartParsingException($"Invalid byte value for start of JSON element. " +
-                                                             $"Found = {_buffer[_current]}, " +
+                                                             $"Found = {(char)Current!}, " +
                                                              $"0-Based Position = {Position}.")
             };
         }
 
-        private JsonType SkipArray()
+        private JsonType SkipArrayAsync(CancellationToken token)
         {
-            if (NextWithEnsureCapacity())
+            if (_buffer.MoveNextAsync(token))
             {
-                SkipWhiteSpaceWithVerify("]");
-                while (_buffer[_current] != JsonConst.ArrayEndByte)
+                SkipWhiteSpaceWithVerifyAsync("]", token);
+                while (_buffer.Current != JsonConst.ArrayEndByte)
                 {
-                    SkipUntilNextRaw();
-                    ReadIsValueSeparationOrEndWithVerify(JsonConst.ArrayEndByte,
+                    SkipUntilNextRawAsync(token);
+                    ReadIsValueSeparationOrEndWithVerifyAsync(JsonConst.ArrayEndByte,
                             "array",
                             "',' or ']' (but not ',]')",
-                            default);
+                            token);
                 }
-                NextWithEnsureCapacity();
+                _buffer.MoveNextAsync(token);
                 return JsonType.Arr;
             }
             throw new JsonArrayPartParsingException($"Reached end, unable to find valid JSON end-array (']'). " +
-                                                    $"0-Based Position = {Position}.");
+                                           $"0-Based Position = {Position}.");
         }
 
-        private JsonType SkipObject()
+        private JsonType SkipObjectAsync(CancellationToken token)
         {
-            if (NextWithEnsureCapacity())
+            if (_buffer.MoveNextAsync(token))
             {
-                SkipWhiteSpaceWithVerify("}");
-                while (_buffer[_current] != JsonConst.ObjectEndByte)
+                SkipWhiteSpaceWithVerifyAsync("}", token);
+                while (_buffer.Current != JsonConst.ObjectEndByte)
                 {
-                    if (_buffer[_current] != JsonConst.StringQuoteByte)
+                    if (_buffer.Current != JsonConst.StringQuoteByte)
                     {
                         throw new JsonArrayPartParsingException($"Invalid byte value for start of Object Property Name. " +
-                                                                $"Expected = {JsonConst.StringQuoteByte}, " +
-                                                                $"Found = {_buffer[_current]}, " +
-                                                                $"0-Based Position = {Position}.");
+                            $"Expected = {JsonConst.StringQuoteByte}, " +
+                            $"Found = {(char)Current!}, " +
+                            $"0-Based Position = {Position}.");
                     }
-                    SkipString();
-                    SkipWhiteSpaceWithVerify(":");
-                    _current--;
-                    NextExpectedOrThrow(JsonConst.NameSeparatorByte, "Object property");
-                    _current++;
-                    SkipWhiteSpaceWithVerify("Object property value");
-                    SkipUntilNextRaw();
-                    ReadIsValueSeparationOrEndWithVerify(JsonConst.ObjectEndByte,
+                    SkipStringAsync(token);
+                    SkipWhiteSpaceWithVerifyAsync(":", token);
+                    _buffer.StepBack();
+                    NextExpectedOrThrowAsync(JsonConst.NameSeparatorByte, token, "Object property");
+                    _buffer.MoveNextAsync(token);
+                    SkipWhiteSpaceWithVerifyAsync("Object property value", token);
+                    SkipUntilNextRawAsync(token);
+                    ReadIsValueSeparationOrEndWithVerifyAsync(JsonConst.ObjectEndByte,
                             "Object property",
-                            "',' or '}' (but not ',}')", 
-                            default);
+                            "',' or '}' (but not ',}')",
+                            token);
                 }
-                NextWithEnsureCapacity();
+                _buffer.MoveNextAsync(token);
                 return JsonType.Obj;
             }
             throw new JsonArrayPartParsingException($"Reached end, unable to find valid JSON end-object ('}}'). " +
-                                                    $"0-Based Position = {Position}.");
+                                           $"0-Based Position = {Position}.");
         }
 
-        private JsonType SkipString()
+        private void ReadIsValueSeparationOrEndWithVerifyAsync(byte end, string partOf, string expected, CancellationToken token)
         {
-            while (NextWithEnsureCapacity())
+            if (ReadIsValueSeparationOrEndAsync(end, token)) return;
+            if (InRange)
             {
-                switch (_buffer[_current])
+                throw new JsonArrayPartParsingException($"Invalid byte value for '{partOf}'. " +
+                    $"Expected = {expected}, " +
+                    $"Found = {(char)Current!}, " +
+                    $"0-Based Position = {Position}.");
+            }
+            throw new JsonArrayPartParsingException($"Reached end, unable to find '{end}'. " +
+                                           $"0-Based Position = {Position}.");
+        }
+
+        private bool ReadIsValueSeparationOrEndAsync(byte end, CancellationToken token)
+        {
+            SkipWhiteSpaceAsync(token);
+            if (!InRange) return false;
+            if (_buffer.Current == end) return true;
+            if (_buffer.Current != JsonConst.ValueSeparatorByte) return false;
+            _buffer.MoveNextAsync(token);
+            SkipWhiteSpaceAsync(token);
+            return InRange && _buffer.Current != end;
+        }
+
+        private JsonType SkipStringAsync(CancellationToken token)
+        {
+            while (_buffer.MoveNextAsync(token))
+            {
+                switch (_buffer.Current)
                 {
                     case JsonConst.ReverseSlashByte:
-                        if (NextWithEnsureCapacity())
+                        if (_buffer.MoveNextAsync(token))
                         {
-                            switch (_buffer[_current])
+                            switch (_buffer.Current)
                             {
                                 case JsonConst.ReverseSlashByte:
                                 case JsonConst.ForwardSlashByte:
@@ -288,7 +312,7 @@ namespace DevFast.Net.Text.Json.Utf8
                                 case JsonConst.SecondOfHexDigitInStringByte:
                                     //JsonSerializer must handle validation!
                                     //we skip 4 bytes
-                                    if (NextWithEnsureCapacity(4))
+                                    if (_buffer.MoveNextAsync(token, 4))
                                     {
                                         continue;
                                     }
@@ -305,14 +329,14 @@ namespace DevFast.Net.Text.Json.Utf8
                                         $"\\{JsonConst.FirstOfTrueByte} or " +
                                         $"\\{JsonConst.LastOfCarriageReturnInStringByte} or " +
                                         $"\\{JsonConst.SecondOfHexDigitInStringByte}4Hex, " +
-                                        $"Found = \\{_buffer[_current]}, " +
+                                        $"Found = \\{(char)Current!}, " +
                                         $"0-Based Position = {Position}.");
                             }
                         }
                         throw new JsonArrayPartParsingException($"Reached end, unable to find valid escape character. " +
                                                        $"0-Based Position = {Position}.");
                     case JsonConst.StringQuoteByte:
-                        NextWithEnsureCapacity();
+                        _buffer.MoveNextAsync(token);
                         return JsonType.Str;
                 }
             }
@@ -320,14 +344,14 @@ namespace DevFast.Net.Text.Json.Utf8
                                            $"0-Based Position = {Position}.");
         }
 
-        private JsonType SkipNumberWithoutValidation()
+        private JsonType SkipNumberWithoutValidationAsync(CancellationToken token)
         {
             //we just take everything until begin of next token (even if number is not valid!)
             //number parsing rules are too much to write here
             //Serializer will do its job during deserialization
-            while (NextWithEnsureCapacity())
+            while (_buffer.MoveNextAsync(token))
             {
-                switch (_buffer[_current])
+                switch (_buffer.Current)
                 {
                     case JsonConst.MinusSignByte:
                     case JsonConst.PlusSignByte:
@@ -351,163 +375,137 @@ namespace DevFast.Net.Text.Json.Utf8
             return JsonType.Num;
         }
 
-        private JsonType SkipRueOfTrueWithoutValidation()
+        private JsonType SkipRueOfTrueWithoutValidationAsync(CancellationToken token)
         {
             //JsonSerializer must handle literal validation!
             //we skip 3 bytes
-            if (NextWithEnsureCapacity(3))
+            if (_buffer.MoveNextAsync(token, 3))
             {
                 //this one to move the pointer forward, we don't care
                 //about EoF, that's handled by next read!
-                NextWithEnsureCapacity();
+                _buffer.MoveNextAsync(token);
                 return JsonType.Bool;
             }
             throw new JsonArrayPartParsingException($"Reached end while parsing 'true' literal. " +
                                                     $"0-Based Position = {Position}.");
         }
 
-        private JsonType SkipAlseOfFalseWithoutValidation()
+        private JsonType SkipAlseOfFalseWithoutValidationAsync(CancellationToken token)
         {
             //JsonSerializer must handle literal validation!
             //we skip 4 bytes
-            if (NextWithEnsureCapacity(4))
+            if (_buffer.MoveNextAsync(token, 4))
             {
                 //this one to move the pointer forward, we don't care
                 //about EoF, that's handled by next read!
-                NextWithEnsureCapacity();
+                _buffer.MoveNextAsync(token);
                 return JsonType.Bool;
             }
             throw new JsonArrayPartParsingException($"Reached end while parsing 'false' literal. " +
                                                     $"0-Based Position = {Position}.");
         }
 
-        private JsonType SkipUllOfNullWithoutValidation()
+        private JsonType SkipUllOfNullWithoutValidationAsync(CancellationToken token)
         {
             //JsonSerializer must handle literal validation!
             //we skip 3 bytes
-            if (NextWithEnsureCapacity(3))
+            if (_buffer.MoveNextAsync(token, 3))
             {
                 //this one to move the pointer forward, we don't care
                 //about EoF, that's handled by next read!
-                NextWithEnsureCapacity();
+                _buffer.MoveNextAsync(token);
                 return JsonType.Null;
             }
             throw new JsonArrayPartParsingException($"Reached end while parsing 'null' literal. " +
                                                     $"0-Based Position = {Position}.");
         }
 
-        private void ReadIsValueSeparationOrEndWithVerify(byte end, string partOf, string expected, CancellationToken token)
+        private void NextExpectedOrThrowAsync(byte expected, CancellationToken token, string partOf)
         {
-            if (ReadIsValueSeparationOrEnd(end, token)) return;
-            if (InRange)
-            {
-                throw new JsonArrayPartParsingException($"Invalid byte value for '{partOf}'. " +
-                                                        $"Expected = {expected}, " +
-                                                        $"Found = {_buffer[_current]}, " +
-                                                        $"0-Based Position = {Position}.");
-            }
-            throw new JsonArrayPartParsingException($"Reached end, unable to find '{end}'. " +
-                                                    $"0-Based Position = {Position}.");
-        }
-
-        private bool ReadIsValueSeparationOrEnd(byte end, CancellationToken token)
-        {
-            SkipWhiteSpace();
-            if (!InRange) return false;
-            if (_buffer[_current] == end) return true;
-            if (_buffer[_current] != JsonConst.ValueSeparatorByte) return false;
-            _current++;
-            SkipWhiteSpace();
-            token.ThrowIfCancellationRequested();
-            return InRange && _buffer[_current] != end;
-        }
-
-        private bool ReadIsGivenByte(byte match, CancellationToken token)
-        {
-            SkipWhiteSpace();
-            token.ThrowIfCancellationRequested();
-            if (!InRange || _buffer[_current] != match) return false;
-            _current++;
-            return true;
-        }
-
-        private void SkipWhiteSpaceWithVerify(string jsonToken)
-        {
-            SkipWhiteSpace();
-            if (!InRange)
-                throw new JsonArrayPartParsingException($"Reached end, expected to find '{jsonToken}'. " +
-                                                        $"0-Based Position = {Position}.");
-        }
-
-        private void NextExpectedOrThrow(byte expected, string partOf)
-        {
-            if (NextWithEnsureCapacity() && _buffer[_current] == expected) return;
+            if (_buffer.MoveNextAsync(token) && _buffer.Current == expected) return;
             if (InRange)
             {
                 throw new JsonArrayPartParsingException($"Invalid byte value while parsing '{partOf}'. " +
-                                                        $"Expected = {expected}, " +
-                                                        $"Found = {_buffer[_current]}, " +
-                                                        $"0-Based Position = {Position}.");
+                                               $"Expected = {expected}, " +
+                                               $"Found = {(char)Current!}, " +
+                                               $"0-Based Position = {Position}.");
             }
             throw new JsonArrayPartParsingException($"Reached end while parsing '{partOf}'. " +
-                                                    $"Expected = {expected}, " +
-                                                    $"0-Based Position = {Position}.");
+                                           $"Expected = {expected}, " +
+                                           $"0-Based Position = {Position}.");
         }
 
-        private void SkipWhiteSpace()
+        private bool ReadIsGivenByteAsync(byte match, CancellationToken token)
+        {
+            SkipWhiteSpaceAsync(token);
+            if (!InRange || _buffer.Current != match) return false;
+            _buffer.MoveNextAsync(token);
+            _buffer.SkipUntilCurrent();
+            return true;
+        }
+
+        private void SkipWhiteSpaceWithVerifyAsync(string jsonToken, CancellationToken token)
+        {
+            SkipWhiteSpaceAsync(token);
+            if (!InRange)
+                throw new JsonArrayPartParsingException($"Reached end, expected to find '{jsonToken}'. " +
+                                               $"0-Based Position = {Position}.");
+        }
+
+        private void SkipWhiteSpaceAsync(CancellationToken token)
         {
             while (InRange)
             {
-                switch (_buffer[_current])
+                switch (_buffer.Current)
                 {
                     case JsonConst.SpaceByte:
                     case JsonConst.HorizontalTabByte:
                     case JsonConst.NewLineByte:
                     case JsonConst.CarriageReturnByte:
-                        NextWithEnsureCapacity();
+                        _buffer.MoveNextAsync(token);
                         continue;
                     case JsonConst.ForwardSlashByte:
-                        if (!NextWithEnsureCapacity())
+                        if (!_buffer.MoveNextAsync(token))
                         {
                             throw new JsonArrayPartParsingException("Reached end. " +
                                                            "Can not find correct comment format " +
                                                            "(neither single line comment token '//' " +
                                                            "nor multi-line comment token '/*').");
                         }
-                        ReadComment();
+                        ReadCommentAsync(token);
                         continue;
                     default: return;
                 }
             }
         }
 
-        private void ReadComment()
+        private void ReadCommentAsync(CancellationToken token)
         {
-            switch (_buffer[_current])
+            switch (_buffer.Current)
             {
                 case JsonConst.ForwardSlashByte:
-                    while (NextWithEnsureCapacity())
+                    while (_buffer.MoveNextAsync(token))
                     {
-                        if (_buffer[_current] != JsonConst.CarriageReturnByte &&
-                            _buffer[_current] != JsonConst.NewLineByte) continue;
-                        NextWithEnsureCapacity();
+                        var current = _buffer.Current;
+                        if (current != JsonConst.CarriageReturnByte && current != JsonConst.NewLineByte) continue;
+                        _buffer.MoveNextAsync(token);
                         return;
                     }
                     //we don't throw if we reach EoJ, we consider comment ended there!
                     //so we get out. If any other token was expected, further parsing will throw
                     //proper error.
-                    break;  
+                    break;
                 case JsonConst.AsteriskByte:
-                    while (NextWithEnsureCapacity())
+                    while (_buffer.MoveNextAsync(token))
                     {
-                        if (_buffer[_current] != JsonConst.AsteriskByte) continue;
-                        if (NextWithEnsureCapacity() &&
-                            _buffer[_current] == JsonConst.ForwardSlashByte)
+                        if (_buffer.Current != JsonConst.AsteriskByte) continue;
+                        if (_buffer.MoveNextAsync(token) &&
+                            _buffer.Current == JsonConst.ForwardSlashByte)
                         {
-                            NextWithEnsureCapacity();
+                            _buffer.MoveNextAsync(token);
                             return;
                         }
-                        _current--;
+                        _buffer.StepBack();
                     }
                     //we need to throw error even if we reached EoJ
                     //coz the comment was not properly terminated!
@@ -521,20 +519,58 @@ namespace DevFast.Net.Text.Json.Utf8
             }
         }
 
-        private bool NextWithEnsureCapacity(int count = 1)
+        private void ReDefineBuffer(int offsetIncrement, CancellationToken token)
         {
-            _current += count;
-            return InRange;
+            IncreaseConsumption(offsetIncrement);
+            if (_stream == null) return;
+            if (_current >= (_buffer.Length + 1) / 2)
+            {
+                _end -= _current;
+                if (_end > 0) _buffer.LiftNCopyUnSafe(_current, _end, 0);
+                _current = _begin = 0;
+            }
+            if (_end < _buffer.Length) FillBuffer(token);
+        }
+
+        private bool TryIncreasingBuffer(CancellationToken token)
+        {
+            if (_stream == null) return false;
+            if (_end == _buffer.Length) _buffer = _buffer.DoubleByteCapacity();
+            return FillBuffer(token);
+        }
+
+        private bool FillBuffer(CancellationToken token)
+        {
+            var end = _stream!.ReadAsync(_buffer.AsMemory(_end, _buffer.Length - _end), token).AsTask().GetAwaiter().GetResult();
+            if (end == 0)
+            {
+                DisposeStream();
+                return false;
+            }
+
+            _end += end;
+            return true;
+        }
+
+        private void IncreaseConsumption(int offsetIncrement)
+        {
+            _current += offsetIncrement;
+            _bytesConsumed += (_current - _begin);
+            _begin = _current;
         }
 
         /// <summary>
-        /// Asynchronous clean up by releasing resources.
+        /// Clean up by releasing resources.
         /// </summary>
         public void Dispose()
         {
+            DisposeStream();
+        }
+
+        private void DisposeStream()
+        {
             if (_stream != null && _disposeStream) _stream.Dispose();
             _stream = null;
-            _buffer = Array.Empty<byte>();
         }
     }
 }
